UDP inputs now uses human friendly size to define MaxMessageSize (#6886)
Uses go-humanize to parse values for `MaxMessageSize` and fallback to
raw bytes if no suffix is found. Also create a new type for future
configuration named `cfgtype.ByteSize` that will take care of correctly
unpacking values.

+++ filebeat/inputsource/tcp/server.go
@@ -89,7 +89,7 @@ func (s *Server) run() {
 			s.log,
 			s.callback,
 			s.splitFunc,
-			s.config.MaxMessageSize,
+			uint64(s.config.MaxMessageSize),
 			s.config.Timeout,
 		)
